+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Layout from '../components/Layout/Layout'
 
-import Header from '../components/Header'
+import Header from '../components/Header/Header'
 import Main from '../components/Main'
 import Footer from '../components/Footer/Footer'
